+++ print-icmp6.c
@@ -776,7 +776,7 @@ print_lladdr(netdissect_options *ndo, const uint8_t *p, size_t l)
 	}
 }
 
-static int icmp6_cksum(netdissect_options *ndo, const struct ip6_hdr *ip6,
+static uint16_t icmp6_cksum(netdissect_options *ndo, const struct ip6_hdr *ip6,
 	const struct icmp6_hdr *icp, u_int len)
 {
 	return nextproto6_cksum(ndo, ip6, (const uint8_t *)(const void *)icp, len, len,
